// frontend/lib/api/models/RedirectResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Redirect response.
 */
export type RedirectResponse = {
    id: string;
    source_path: string;
    target_path: string;
    status_code: number;
    enabled: boolean;
    created_at: string;
    updated_at: string;
    notes?: (string | null);
};